Change check for climo years
This will now check if climo start and end years care not specified in the config file.

If not, it uses the initial case_name var. If the years are specified, it will append to the case_name var.

Prior to this change, the script would fail at the case_name var if the climo years were not specified.

Also add warning to notify user just in case