Change check for missing climo years
This will change the check from whether the climo years exist to if they are NoneType.

The previous check failed at the case_name var in the perform_analyses function if the climo years didn't exist with a NoneType error.

Also add a warning message for the user just in case.