Test to try and stay ahead of polygon failures in the plot
This is meant to try and fix the percent difference plots in the polar plots. With out a correction the plots are wrong, and with the correction sometimes the plotting fails due to polygon errors. This will allow the correct plots, but if there is a polygon error, skip that particular plot.  This is a test and by all means not the final solution!