Use xr.DataArray.data as arguments to np functions
Some combinations of recent versions of xarray and numpy throw a ValueError
when the first two arguments to np.linspace are DataArray objects instead of
numpy.array objects. This can be avoided by using the .data component as the
argument.

I've verified that in older numpy versions where the xarray argument is
allowed, the result from np.linspace is bit-for-bit identical whether the
argument is the DataArray or DataArray.data